GCC warning fix

+++ Render/src/Resources/VolatileMaterialValues.c
@@ -53,7 +53,7 @@ static uint32_t identityHash(const void* key)
 static unsigned int getTableSize(unsigned int maxValues)
 {
 	const float loadFactor = 0.75f;
-	return (unsigned int)(maxValues/loadFactor);
+	return (unsigned int)((float)maxValues/loadFactor);
 }
 
 static void* getValue(const dsVolatileMaterialValues* values, uint32_t nameId, Type type)
